+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## [WIP]
 
+- Add version string to settings
 - Add Puffynetwork [#82]
 - Add PlayboyPlus [#79],[#80]
 - Improve Hegre [#79]